requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# OCR and Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.4.0
opencv-python-headless==4.10.0.84
PyPDF2==3.0.1

# LLM Integration - Using Groq (FREE!)
groq==0.4.1

# Image preprocessing
numpy==1.26.4
scikit-image==0.24.0

# URL handling
httpx==0.27.2

# Utilities
requests==2.31.0
aiofiles==23.2.1
